fixes and test refactoring
